app/requirements.txt
Flask~=2.3.2
keras~=3.3.3
tensorflow~=2.16.1
numpy~=1.23.5
Pillow~=9.5.0
gevent~=22.10.2
opencv-python ~= 4.8.0